Fix Order entry
The ProfitTaker and StopLoss were previously only submitted as
separate entries, and thus weren't triggering since they were not
connected to the parent entry. "BracketOrder" was just a data
structure that doesn't do anything with the entries.

I copied the ib.py bracketOrder code to properly create the orders,
I also added on an additional market sell order to trigger at end
of day if neither the StopLimit nor the ProfitTaker are reached